+++ src/Mpociot/BotMan/Drivers/FacebookDriver.php
@@ -62,6 +62,23 @@ class FacebookDriver extends Driver
     }
 
     /**
+     * @param Message $matchingMessage
+     * @return void
+     */
+    public function types(Message $matchingMessage)
+    {
+        $parameters = [
+            'recipient' => [
+                'id' => $matchingMessage->getChannel(),
+            ],
+            'access_token' => $this->config->get('facebook_token'),
+            'sender_action' => 'typing_on',
+        ];
+
+        return $this->http->post('https://graph.facebook.com/v2.6/me/messages', [], $parameters);
+    }
+
+    /**
      * @param  Message $message
      *
      * @return Answer
@@ -87,7 +104,7 @@ class FacebookDriver extends Driver
     {
         $messages = Collection::make($this->event->get('messaging'));
         $messages = $messages->transform(function ($msg) {
-            if (isset($msg['message'])) {
+            if (isset($msg['message']) && isset($msg['message']['text'])) {
                 return new Message($msg['message']['text'], $msg['recipient']['id'], $msg['sender']['id'], $msg);
             } elseif (isset($msg['postback'])) {
                 return new Message($msg['postback']['payload'], $msg['recipient']['id'], $msg['sender']['id'], $msg);