+++ src/Mpociot/BotMan/Traits/VerifiesServices.php
@@ -11,10 +11,11 @@ trait VerifiesServices
     /**
      * Verify service webhook URLs.
      *
-     * @param string $facebookVerification the Facebook verification string to match
+     * @param string $facebookVerification The Facebook verification string to match
+     * @param string $weChatVerification The WeChat verification token to match
      * @return Response
      */
-    public function verifyServices($facebookVerification)
+    public function verifyServices($facebookVerification = null, $weChatVerification = null)
     {
         $request = Request::createFromGlobals();
         $payload = Collection::make(json_decode($request->getContent(), true));
@@ -27,6 +28,18 @@ trait VerifiesServices
         // Facebook verification
         if ($request->get('hub_mode') === 'subscribe' && $request->get('hub_verify_token') === $facebookVerification) {
             return Response::create($request->get('hub_challenge'))->send();
+        }
+
+        // WeChat verification
+        if ($request->get('signature') !== null && $request->get('timestamp') !== null && $request->get('nonce') !== null && $request->get('echostr') !== null) {
+            $tmpArr = [$weChatVerification, $request->get('timestamp'), $request->get('nonce')];
+            sort($tmpArr, SORT_STRING);
+            $tmpStr = implode($tmpArr);
+            $tmpStr = sha1($tmpStr);
+
+            if ($tmpStr == $request->get('signature')) {
+                return Response::create($request->get('echostr'))->send();
+            }
         }
     }
 }